refactor(action): -d use number replace string

+++ src/utils/time.rs
@@ -26,13 +26,13 @@ pub fn get_now_timestamp() -> Result<i64> {
     Ok(time.timestamp())
 }
 
-pub fn get_date_with_offset(fmt: &str, day: i64) -> String {
+pub fn get_date_with_offset(fmt: &str, day: u32) -> String {
     // Time zone offset
     let offset: chrono::FixedOffset =
         FixedOffset::east_opt(8 * 3600).expect("time zone offset fail");
 
     let dt = Utc::now().timestamp();
-    let n_day = 60 * 60 * 24 * day;
+    let n_day = 60 * 60 * 24 * day as i64;
     let now = dt + n_day;
     let local = Utc.timestamp_opt(now, 0);
     let date = match local {
